requirements.txt
streamlit==1.5.0
Pillow==8.3.2 
